refactor(ipc): hide `PipeServer` implemention
