+++ include/WeaselIPC.h
@@ -31,8 +31,6 @@ enum WEASEL_IPC_COMMAND
 
 namespace weasel
 {
-	// 为了 32 位 server 和 64 位 TSF 的兼容，这里不用 LPARAM 和 WPARAM
-	// ServerImpl 处也有一样的问题。
 	struct PipeMessage {
 		WEASEL_IPC_COMMAND Msg;
 		UINT wParam;
